Python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "GnG"
version = "0.0.1"
authors = [
  { name="Eric", email="171273411+Cnecomicz@users.noreply.github.com" }
]
description = "Gold & Gallows TODO"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
  "flexmock",
  "pylint",
  "pytest",
  "pytest-cov",
  "python-statemachine"
]

[template.plugins.default]
src-layout = true

[project.urls]
Homepage = "https://github.com/cnecomicz/python-rpg-game"
Issues = "https://github.com/cnecomicz/python-rpg-game/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/GnG"]

[project.scripts]
gng = "GnG.main:main"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
